+++ src/main/java/pro/sky/java/course2/exception/workwithexceptionproject/Employee.java
@@ -1,4 +1,5 @@
 package pro.sky.java.course2.exception.workwithexceptionproject;
+
 import java.util.Objects;
 
 public class Employee {
