+++ src/main/java/pro/sky/java/course2/exception/workwithexceptionproject/service/EmployeeService.java
@@ -1,53 +1,11 @@
 package pro.sky.java.course2.exception.workwithexceptionproject.service;
 
-import org.springframework.stereotype.Service;
-import pro.sky.java.course2.exception.workwithexceptionproject.ArraysIsFullException;
 import pro.sky.java.course2.exception.workwithexceptionproject.Employee;
-import pro.sky.java.course2.exception.workwithexceptionproject.EmployeeIsNotFoundException;
 
-@Service
-public class EmployeeService {
-    private final Employee[] employees = new Employee[1];
+public interface EmployeeService {
+    Employee addEmployee(String firstName, String lastName);
 
+    Employee removeEmployee(String firstName, String lastName);
 
-    public void addEmployee(String firstName, String lastName) {
-        for (int i = 0; i < employees.length; i++) {
-            if (employees[i] == null) {
-                employees[i] = new Employee(firstName, lastName);
-                return;
-            }
-        }
-        throw new ArraysIsFullException();
-    }
-
-    public void removeEmployee(String firstName, String lastName) {
-        for (int i = 0; i < employees.length; i++) {
-            if (employees[i] == null) {
-                continue;
-            }
-            if (employees[i].getFirstName().equals(firstName) && employees[i].getLastName().equals(lastName)) {
-                employees[i] = null;
-                return;
-            }
-        }
-        throw new EmployeeIsNotFoundException();
-    }
-
-    public Employee findEmployee(String firstName, String lastName) {
-        for (Employee employee : employees) {
-            if (employee == null) {
-                continue;
-            }
-            if (firstName.equals(employee.getFirstName()) && lastName.equals(employee.getLastName())) {
-                return employee;
-            }
-        }
-        throw new EmployeeIsNotFoundException();
-    }
+    Employee findEmployee(String firstName, String lastName);
 }
-
-
-
-
-
-
